allow expiration field for passworless to support duration

+++ support/cas-server-support-passwordless-api/src/main/java/org/apereo/cas/impl/token/InMemoryPasswordlessTokenRepository.java
@@ -23,7 +23,7 @@ public class InMemoryPasswordlessTokenRepository extends BasePasswordlessTokenRe
 
     private final Cache<String, String> storage;
 
-    public InMemoryPasswordlessTokenRepository(final int tokenExpirationInSeconds,
+    public InMemoryPasswordlessTokenRepository(final long tokenExpirationInSeconds,
                                                final CipherExecutor cipherExecutor) {
         super(tokenExpirationInSeconds, cipherExecutor);
         storage = Caffeine.newBuilder()
